Simplified and Generalized Validation Handling UI Side
Will now have a standard way of handling mutations
