+++ client/src/components/admin/characterList/UpdateGoFields.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
+import toaster from '@/services/Toasters'
 
-import { inject, ref, type Ref, watch } from 'vue'
+import { inject, ref, type Ref } from 'vue'
 import FormInputNumberWrapper from '@/FormWrappers/FormInputNumberWrapper.vue'
 import Button from 'primevue/button'
 import { getValidationInstance } from '@/components/admin/characterList/validators/characterGoFieldsForm.ts'
@@ -8,54 +9,30 @@ import FormWrapper from '@/FormWrappers/FormWrapper.vue'
 import { useQuery } from '@pinia/colada'
 import { goFieldQuery, useUpdateGoFields } from '@/components/admin/characterList/stores/goFieldColada.ts'
 import type { GoFields } from '@/components/admin/characterList/types.ts'
+import { useHydrateFormOnce } from '@/utilities/piniaColadaUtilities.ts'
 
 const form = getValidationInstance()
 const dialogRef = inject('dialogRef') as Ref
 const characterId = ref(dialogRef.value.data.characterId as number)
 
 const { data, isPending } = useQuery(goFieldQuery(characterId.value))
-const updateGoFields = useUpdateGoFields()
-
 useHydrateFormOnce(data, form.setValues)
+
+const updateGoFields = useUpdateGoFields((errors) => {
+  form.setErrors(errors)
+})
 
 const onSubmit = form.handleSubmit(async (values) => {
   await updateGoFields.mutateAsync({
     id: characterId.value,
-    data: {
-      wealthLevel: values.wealthLevel,
-      motes: values.motes,
-      primaFragments: values.primaFragments,
-      voidFragments: values.voidFragments,
-    } as GoFields })
+    data: values as GoFields,
+  })
+  toaster.success('Go fields updated successfully')
   cancel()
 })
 
 const cancel = () => {
   dialogRef.value.close()
-}
-
-function useHydrateFormOnce<T>(
-  source: Ref<T | undefined>,
-  hydrate: (value: T) => void,
-) {
-  const hydrated = ref(false)
-
-  watch(
-    source,
-    (val) => {
-      if (!val || hydrated.value) return
-
-      hydrate(val)
-      hydrated.value = true
-    },
-    { immediate: true },
-  )
-
-  return {
-    resetHydration: () => {
-      hydrated.value = false
-    },
-  }
 }
 
 </script>
