+++ client/src/components/admin/characterList/services/goFieldsService.ts
@@ -5,5 +5,5 @@ export const goFieldsService = {
   getGoFields: (characterId: number): Promise<GoFields> => axios.get<GoFields>(`/characters/${characterId}/goFields`)
     .then(async (response) => { return response.data }),
   updateGoFields: (characterId: number, goFields: GoFields) => axios.put(`/characters/${characterId}/goFields`, goFields)
-    .then(async (response) => { return response.data }),
+    .then(async (response) => { return response }),
 }
