+++ client/src/components/admin/characterList/stores/goFieldColada.ts
@@ -1,7 +1,7 @@
 import { defineQuery, useMutation, useQueryCache } from '@pinia/colada'
 
-import toaster from '@/services/Toasters'
 import { goFieldsService } from '@/components/admin/characterList/services/goFieldsService.ts'
+import { handleValidationErrors } from '@/utilities/piniaColadaUtilities.ts'
 import type { GoFields } from '@/components/admin/characterList/types.ts'
 
 export const GO_FIELDS_QUERY_KEYS = {
@@ -14,14 +14,16 @@ export const goFieldQuery = (characterId: number) =>
     query: () => goFieldsService.getGoFields(characterId),
   }))
 
-export const useUpdateGoFields = () => {
+export const useUpdateGoFields = (onValidationError?: (errors: Record<string, any>) => void | undefined) => {
   const queryCache = useQueryCache()
 
   return useMutation({
     mutation: ({ id, data }: { id: number, data: GoFields }) => goFieldsService.updateGoFields(id, data),
     async onSuccess() {
-      toaster.success('Successfully Updated Go Fields!')
       await queryCache.invalidateQueries({ key: GO_FIELDS_QUERY_KEYS.root })
+    },
+    onError(error: any) {
+      handleValidationErrors(error, onValidationError)
     },
   })
 }
